replaced, more optimised algorithm

+++ input_processing/src/main.rs
@@ -99,7 +99,7 @@ fn main() {
     for sample_vector in dataset_vector {
         let delta_layer: Vec<i32> = Vec::new();
         for pixel_pointer in 0..=sample_vector.len() {
-            delta_layer.push(sample_vector[pixel_pointer] - input_image_linear_vector[pixel_pointer]);
+            delta_layer.push(sample_vector[pixel_pointer] - input_image_linear_vector[pixel_pointer] as i32);
         }
         delta_vector.push(delta_layer);
     }
